Group 5/group05_version1/requirements.txt
pytorch
torchvision
numpy
tqdm
timm
wandb
einops
cv2
matplotlib
